update dockerfile and appbuild to use ninja

+++ appbuild.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
 
-rm -rf paragon_back
+if [ -d "paragon_back" ]; then
+    rm -rf paragon_back
+fi
 
-git clone -b $BRANCH  $REPO
+git clone -b $BRANCH $REPO paragon_back
 cd paragon_back
 mkdir build
 cd build
-cmake ..
+cmake -GNinja ..
 cd ..
-cmake --build build
-cd build
-ctest
-cd ../..
+cmake --build build --target all
+ctest --test-dir build
+cd ..
